--- src/lib.rs
@@ -1,2 +1,0 @@
-pub mod cmd;
-pub mod input_processor;
